requirements/dev.txt
black==23.9.1
django-debug-toolbar==4.2.0